+++ lib/main.dart
@@ -1,27 +1,39 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_deep_state_manage/presentation/controllers/theme_controller.dart';
-import 'package:flutter_deep_state_manage/presentation/pages/my_home_page.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
-final themeController = ThemeController();
-
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
+  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return ValueListenableBuilder(
-      valueListenable: themeController,
-      builder: (context, value, child) {
-        return MaterialApp(
-          title: 'State Manage',
-          theme: value ? ThemeData.dark() : ThemeData.light(),
-          home: const MyHomePage(title: 'Flutter State Manage'),
-        );
-      },
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+      ),
+      home: const MyHomePage(),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key});
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text('Flutter State Management'),
+      ),
     );
   }
 }
