--- lib/presentation/controllers/theme_controller.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ThemeController extends ValueNotifier<bool> {
-  ThemeController() : super(false);
-
-  void toggleTheme() {
-    value = !value;
-    notifyListeners();
-  }
-}
