--- lib/presentation/pages/my_home_page.dart
@@ -1,43 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../main.dart';
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const Text('Trocar tema do aplicativo'),
-            ValueListenableBuilder(
-              valueListenable: themeController,
-              builder: (context, value, child) {
-                return Switch(
-                  value: value,
-                  onChanged: (value) {
-                    themeController.toggleTheme();
-                  },
-                );
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
